move UserImage to components, update imports
