+++ src/features/messages/UserMessageDetail.js
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom'
 import { saveState } from '../../helpers/localStorage'
 import { TimeAgo } from '../../components/TimeAgo'
 import useVisualMode from '../../hooks/useVisualMode'
-import UserImage from '../users/UserImage'
+import UserImage from '../users/components/UserImage'
 import { selectUserById } from '../../reducers/usersSlice'
 import { removeMessage } from '../../reducers/userConversationsSlice'
 import { unwrapResult } from '@reduxjs/toolkit'
