+++ src/features/users/components/UserImage/index.js
@@ -1,0 +1,2 @@
+import UserImage from './UserImage'
+export default UserImage
